web/requirements.txt
fastapi==0.*
uvicorn==0.*
asyncpg==0.*
python-dotenv==1.*